fix warnings

+++ include/SWI-Prolog.h
@@ -114,7 +114,7 @@ typedef	void *record_t;
 typedef uintptr_t	atom_t;
 typedef	YAP_Term    *predicate_t;
 typedef struct  open_query_struct *qid_t;
-typedef long    functor_t;
+typedef uintptr_t    functor_t;
 typedef int     (*PL_agc_hook_t)(atom_t);
 typedef unsigned long	foreign_t;	/* return type of foreign functions */
 typedef wchar_t pl_wchar_t;             /* wide character support */
